Update stopped element to match placeholder width
The element that gets pinned to the top of the page is modified to be fixed position, which removes it from the flow of the page and its width is not updated when the page width changes. The placeholder <div> that is inserted into the page will adjust its size when the page size changes, so we can just grab its width when the window size changes and apply it to the pinned element.